[Nu-8818] Let it snow from the root!
