+++ designer/client/src/containers/NussknackerApp.tsx
@@ -3,11 +3,13 @@ import { isEmpty } from "lodash";
 import { HTML5toTouch } from "rdndmb-html5-to-touch";
 import React from "react";
 import { DndProvider } from "react-dnd-multi-backend";
+import Snowfall from "react-snowfall";
 
 import { AiAssistantButton } from "../components/aiAssistant/components/AiAssistantButton";
 import { MenuBar } from "../components/MenuBar";
 import { VersionInfo } from "../components/versionInfo";
 import { getLoggedUser } from "../reducers/selectors/settings";
+import { getUserSettings } from "../reducers/selectors/userSettings";
 import { useAppSelector } from "../store/storeHelpers";
 import { WindowManager } from "../windowManager/WindowManager";
 import { ConnectionErrorProvider } from "./connectionErrorProvider/ConnectionErrorProvider";
@@ -19,6 +21,8 @@ import { useAnonymousStatistics } from "./useAnonymousStatistics";
 
 export function NussknackerApp() {
     const loggedUser = useAppSelector(getLoggedUser);
+    const settings = useAppSelector(getUserSettings);
+    const isSnowing = settings["scenario.isItSnowing"];
 
     useAnonymousStatistics();
     useRegisterTrackingEvents();
@@ -30,6 +34,16 @@ export function NussknackerApp() {
 
     return (
         <>
+            {isSnowing && (
+                <Snowfall
+                    style={{
+                        position: "fixed",
+                        inset: 0,
+                        zIndex: 9999, // higher than JointJS
+                        pointerEvents: "none",
+                    }}
+                />
+            )}
             <DndProvider options={HTML5toTouch}>
                 <WindowManager
                     className={css({
